+++ test/player-test.js
@@ -1,0 +1,1 @@
+const assert = require('chai').assert;
