+++ test/world-test.js
@@ -1,0 +1,16 @@
+const assert = require('chai').assert;
+
+describe('GameWorld', function () {
+
+  context('default', function(){
+
+    it('should be a function', function () {
+      assert.isFunction(GameWorld);
+    });
+    
+  });
+
+
+
+
+});
